tests: Factor fixtures into conftest
Where they should live for other tests to use them.
